walletdk: build the embedded runtime for js/wasm
The embedded daemon path was walled off from js/wasm builds:
embedded.go, embedded_config.go, errmap.go, and the swapruntime /
walletdkrpc tag pairs all carried a !js constraint, so walletdk.Start,
the wallet RPC subserver, and the error-reconstruct interceptor did
not exist in a browser build at all. Drop those constraints so the
in-process bufconn runtime compiles under GOOS=js, and disable the
public HTTP gateway on the embedded path, since a browser has no TCP
listener to bind.

sdk/ark splits its embedded boot into a native file (embedded.go, now
tagged !js || !wasm) and a browser stub (embedded_wasm.go) whose
StartEmbedded reports that the native daemon runtime is unavailable,
with the shared grpc readiness helper lifted into transport.go. The
browser wallet boots the daemon through walletdk.Start rather than
sdk/ark, so the stub keeps the package compiling for js consumers that
only need the remote Ark types.
