+++ sdk/ark/embedded.go
@@ -1,3 +1,5 @@
+//go:build !js || !wasm
+
 package ark
 
 import (
@@ -9,15 +11,8 @@ import (
 	"github.com/lightninglabs/darepo-client/daemonrpc"
 	"github.com/lightninglabs/darepo-client/darepod"
 	"google.golang.org/grpc"
-	"google.golang.org/grpc/connectivity"
 	"google.golang.org/grpc/credentials/insecure"
 	"google.golang.org/grpc/test/bufconn"
-)
-
-const (
-	// defaultBufConnSize is the size of the in-memory listener buffer used
-	// by embedded clients that do not provide their own transport.
-	defaultBufConnSize = 1 << 20
 )
 
 // EmbeddedConfig configures an in-process daemon runtime managed by the SDK.
@@ -244,63 +239,5 @@ func waitForRunExit(ctx context.Context, runErrChan <-chan error) error {
 	case <-ctx.Done():
 		return fmt.Errorf("wait for embedded daemon shutdown: %w",
 			ctx.Err())
-	}
-}
-
-// waitForReady forces a new client connection to attempt dialing and waits
-// until it reaches READY, the embedded daemon exits, or the caller's context
-// expires.
-func waitForReady(ctx context.Context, conn *grpc.ClientConn,
-	runDoneChan <-chan error) error {
-
-	conn.Connect()
-
-	for {
-		state := conn.GetState()
-		if state == connectivity.Ready {
-			return nil
-		}
-
-		if state == connectivity.Shutdown {
-			return fmt.Errorf("grpc connection shut down before " +
-				"readiness")
-		}
-
-		waitCtx, waitCancel := context.WithCancel(ctx)
-		runExitErr := make(chan error, 1)
-		go func() {
-			select {
-			case runErr := <-runDoneChan:
-				if runErr != nil {
-					runExitErr <- fmt.Errorf("embedded "+
-						"daemon exited before "+
-						"readiness: %w", runErr)
-				} else {
-					runExitErr <- fmt.Errorf("embedded " +
-						"daemon exited before " +
-						"readiness")
-				}
-
-				waitCancel()
-
-			case <-waitCtx.Done():
-			}
-		}()
-
-		if !conn.WaitForStateChange(waitCtx, state) {
-			waitCancel()
-
-			select {
-			case err := <-runExitErr:
-				return err
-
-			default:
-			}
-
-			return fmt.Errorf("wait for grpc readiness: %w",
-				ctx.Err())
-		}
-
-		waitCancel()
 	}
 }
