+++ sdk/walletdk/embedded_config.go
@@ -1,5 +1,3 @@
-//go:build !js
-
 package walletdk
 
 import (
