+++ sdk/walletdk/errmap.go
@@ -1,5 +1,3 @@
-//go:build !js
-
 package walletdk
 
 import (
